+++ core/agent_harness/session/integration_resolution.py
